+++ wbstack/CHANGELOG.md
@@ -9,6 +9,7 @@
 - Fix protocol in $wgServer MediaWiki variable
 - Add `wikibaseFedPropsEnable` setting for toggling federated properties
 - Add `wikibaseManifestEquivEntities` setting for WikibaseManifest
+- Set `$wgWBRepoSettings['formatterUrlProperty']` and `$wgWBRepoSettings['canonicalUriProperty']` based on `wikibaseManifestEquivEntities`
 
 ## 1.35-7.4-20210322-2